# place2connect-api/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  handlers.AddUserFriendRequest:
    properties:
      friendID:
        type: integer
      userID:
        type: integer
    required:
    - friendID
    - userID
    type: object
...
  handlers.UpdateUserRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      friends:
        items:
          $ref: '#/definitions/models.User'
        type: array
      impressions:
        type: integer
      isAdmin:
        type: boolean
      isVerified:
        type: boolean
      lastName:
        type: string
      location:
        minLength: 3
        type: string
      occupation:
        minLength: 3
        type: string
      password:
        minLength: 8
        type: string
      photo:
        type: string
      picturePath:
        minLength: 3
        type: string
      provider:
        type: string
      role:
        type: string
      viewedProfile:
        type: integer
    required:
    - firstName
    - lastName
    - location
    - occupation
    - password
    - picturePath
    type: object
  handlers.UserListResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/handlers.UserResponse'
        type: array
      usersCount:
        type: integer
    type: object
  handlers.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      firstName:
        type: string
      friends:
        items:
          $ref: '#/definitions/handlers.UserResponse'
        type: array
      id:
        type: integer
      impressions:
        type: integer
      isAdmin:
        type: boolean
      isVerified:
        type: boolean
      lastName:
        type: string
      location:
        type: string
      occupation:
        type: string
      photo:
        type: string
      picturePath:
        type: string
      provider:
        type: string
      role:
        type: string
      updated_at:
        type: string
      viewedProfile:
        type: integer
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      firstName:
        type: string
      friends:
        items:
          $ref: '#/definitions/models.User'
        type: array
      id:
        type: integer
      impressions:
        type: integer
      isAdmin:
        type: boolean
      lastName:
        type: string
      location:
        type: string
      occupation:
        type: string
      password:
        type: string
      photo:
        type: string
      picturePath:
        type: string
      provider:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      verified:
        type: boolean
      viewedProfile:
        type: integer
    type: object
info:
  contact:
    email: contact@owonafx.com
    name: PLACE2CONNECT API Support
  description: PLACE2CONNECT Docs.
  license:
    name: pace2connect 2.0
  termsOfService: http://swagger.io/terms/
  title: PLACE2CONNECT API
  version: "1.0"
paths:
...
  /api/v1/auth/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a User. Auth is required
      operationId: delete-user
      parameters:
      - description: Id of the user to delete
        in: path
        name: id
        required: true
        type: string
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "404":
          description: Status Not Found
          schema:
            type: string
        "409":
          description: Status Conflict
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a User
      tags:
      - User
  /api/v1/auth/users/{limit}/{offset}:
    get:
      consumes:
      - application/json
      description: Display all Users. Auth is required
      operationId: all-users
      parameters:
      - description: Limit number of products returned (default is 20)
        in: query
        name: limit
        type: integer
      - description: Offset/Skip number of products (default is 0)
        in: query
        name: offset
        type: integer
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserListResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "500":
          description: Status Internal Server Error
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Display all Users
      tags:
      - User
...
  /api/v1/userfriends:
    post:
      consumes:
      - application/json
      description: AddFriendToUser add a new friend to User
      operationId: add-friend
      parameters:
      - description: Add User to Friend
        in: body
        name: userFriend
        required: true
        schema:
          $ref: '#/definitions/handlers.AddUserFriendRequest'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "409":
          description: Status Conflict
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add a new Friend to a User
      tags:
      - User
  /api/v1/userfriends/:
    delete:
      consumes:
      - application/json
      description: RemoveFriendToUser remove a friend to a User
      operationId: remove-friend
      parameters:
      - description: Id of the User to remove the friend
        in: path
        name: userID
        required: true
        type: string
      - description: Id of the UserFriend to remove
        in: path
        name: friendID
        required: true
        type: string
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "409":
          description: Status Conflict
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Remove a new Friend to a User
      tags:
      - User
...
  /api/v1/users/{id}:
    get:
      consumes:
      - application/json
      description: Get a user By ID. Auth required
      operationId: get-user-by-id
      parameters:
      - description: Id of the user to get
        in: path
        name: id
        required: true
        type: integer
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "500":
          description: Status Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a user ID
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update a User. Auth is required
      operationId: update-user
      parameters:
      - description: Id of the user to update
        in: path
        name: id
        required: true
        type: string
      - description: User to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "404":
          description: Status Not Found
          schema:
            type: string
        "409":
          description: Status Conflict
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a user
      tags:
      - User
  /api/v1/users/{userID}/{limit}/{offset}:
    get:
      consumes:
      - application/json
      description: Display all Friends of a User By userId. Auth is required
      operationId: all-user-friends
      parameters:
      - description: ID of the user for friends
        in: path
        name: userID
        required: true
        type: integer
      - description: Limit number of Friends returned (default is 20)
        in: query
        name: limit
        type: integer
      - description: Offset/Skip number of friends (default is 0)
        in: query
        name: offset
        type: integer
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserListResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "500":
          description: Status Internal Server Error
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Display all Friends of a User
      tags:
      - User
  /api/v1/users/email/{email}:
    get:
      consumes:
      - application/json
      description: Get a user. Auth required
      operationId: get-user-by-email
      parameters:
      - description: Email of the user to get
        in: path
        name: email
        required: true
        type: string
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "500":
          description: Status Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a user
      tags:
      - User
  /api/v1/users/me:
    get:
      consumes:
      - application/json
      description: Get current user.
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Logout  of the the API.
      tags:
      - User
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"